Refactor stmt to ease bugfix: reduce duplication
Use the optionMaybe combinator. This reduces code duplication and
left-factorizes the grammar somewhat, potentially improving error
messages. Consider parsing:

foo : * = Pi T . T;

This statement is incorrect, because the grammar requires Pi terms to
have a type annotation, as in `Pi T : * . T`. However, due to
backtracking, we get the wrong error here.

With this fix, when parsing:

foo : * = Pi T . T;

we will no more not try parsing this as `ident args '=' expr` and get a complaint about ": *" (see issue #9).

However, this grammar still requires backtracking, since expressions
can well start with identifiers. So, on the example above, we'll
backtrack and try to parse it as `expr`, and we'll still complain
about the ':'.

Using left-factoring for that seems like a nightmare, subsequent
commits will have a different solution.