Fix #9: Try expressions first, backtrack to definitions
To get the right backtracking behavior (in our case), we need to try
parsing a statement as a definition only after all other alternatives,
that is, after checking if it's a valid expression. This way, if
parsing as a definition fails, we don't backtrack to anything
else.

This might make error messages worse for expressions, but unless
a better fix can be found that's a reasonable tradeoff, because
definitions are often more complex than expressions. As a workaround,
turning `expr;` into `freshName = expr; freshName;` will improve the
error message on `expr`.

However, this would not work without the previous commit: the
beginning of every definition is a syntactically valid expression, so
`try expr` never fails. However, `expr` will not parse the whole
definition; to turn this into a failure, the simplest approach is to
push the parser for the semicolon inside the alternative productions
for statements and inside the try, as done in the previous commit.

Conceptually, pushing a semicolon into the branches should not make a
difference, but that's only valid for CFGs. Even LL(inf) parsers would
have the same problem - (a ';' | b ';') and (a | b) ';' have different
behaviors.

The need for such a transformation points to a modularity problem in
Parsec: you can't give them a compositional semantics mapping them to
CFGs. But the only way to achieve that in general would be using
generalized parsing.