Provide non-monadic wrappers for typecheck*
Also, move other wrappers to PTS.Interactive - the only argument against
that was not copying all the needed imports, but now we need that
anyway.

Additionally, it's good to put these wrappers together because they are
similar and need to be maintained similarly (for instance, they share
the handling of options).