Move runErrorsT into parseCommandLine continuation
This should be safe because parseCommandLine is in CPS and does not use
the Errors monad.

One must also ensure that the Errors monad transformer does nothing
special if the effects are not used; this seems correct by inspection,
but I imagine it would be cooler to establish that using monad
transformer laws.