Expose location of bundled emacs mode via CLI.
This adds a command-line option `--locate-emacs-mode` that prints the
full path of the directory that contains the bundled emacs mode to
stdout.

If pts is installed with `cabal install`, cabal will also install the
emacs mode to some appropriate directory on the target system, and
arrange for `pts --locate-emacs-mode` to return the path of the
installed emacs mode.

If pts is build without installing it, cabal will arrange for `pts
--locate-emacs-mode` to return the path of the emacs mode in the source
tree. I tried this with `cabal run pts` and `cabal repl`.