Implement C-c C-l.
To call pts, the emacs mode needs to know the project directory and
the PTS instance. If there's a module statement in the file, the project
directory is guessed. If that fails, the user is asked once per buffer.

I don't know how to guess the PTS instance, so the user is asked once
per buffer.