Don't evaluate unchecked terms.
This deletes some calls to nbe on the arguments of a call to
typecheckPull.

It is not type-safe to evaluate potentially ill-typed terms. It is not
efficient (to the extent of changing termination behavior) to typecheck
the normal form of a term instead of typechecking the term itself. I
don't know why these calls to nbe were there in the first place.

This increases the exposure of existing bugs where well-typed terms are
compared syntactically instead of comparing up to beta-equivalence.