Remove unit tests about equivalence.
I plan to change equivalence so that it is defined on TypedTerm, but
unit tests on TypedTerm are annoying: For quickcheck properties, we
would need an Arbitrary instance, and for tests, we would need to
annotate the type by hand.

We should cover equivalence with system testing instead, by adding more
examples.