Put more exact Pos annotations.
Before this commit, the Pos annotations of a phrase included all
whitespace or comments after the phrase up till the beginning of the
next phrase in the source file. Now, the Pos annotations only range to
the last character of the phrase. This also fixes a bug with Pos
annotations outside the source file if the next phrase starts on the
first column.