Simplify parseCommandLine by making it non-CPS
parseCommandLine simply calls its argument once in tail position. Then
we can simplify the interface to clarify parseCommandLine simply
computes some value.

In other words, it is in the Cont monad (not even explicitly) but does
not use it for any side effect, because it just calls return for that
monad. Therefore, we can just as well move it out of this monad to make
it more pure.