Remove default language (fomegastar) and hacks
Now, whatever language is selected on the command line cannot be
overriden by files, and fomegastar is no more an exception. It is also
no more the default, so now the language to use will have to be
specified explicitly, either by the main source or on the command line.

This is a regression for existing fomegastar developments, but for
nothing else.