Make pts-quote compile again (not working)
Adapt pts-quote, similarly to pts-nbe in
abbaec64dbc05769c23ca80138e8eaa10689b500. However, here we have a
type-directed transformation, so we need to reify back types to continue
our traversal --- and we need an environment for that!

To make the code compile easily, I used the empty environment; however,
since this is quoting (not normalization), I believe we want an
environment that residualizes every variable (that is, such that `n`
maps to `ResidualVar n` for every used `n`).

However, writing that code without a testcase sounds like a bad idea, so
for now I'll punt.