Halcyon: start post-build-hook for validating sdist
A more complete version is online, but needs to be augmented.

Basic idea: use Halcyon's hash of the source directory, compute the hash
of the result, and compare them. There are some intended differences,
but those can be accounted for by excluding more files from the source
distribution, or by computing instead a new hash for the source.